mutation obserer seems to be observing stuff
Took some time to LITERALLY TEACH MYSELF WHAT A MUTATION OBSERVER IS. finally got it so its actually doing what it need to be doing, next step is trying to interpret what it observes. Contrary to what I thought, the mutationObserver belongs in the content script. Doesn't work in the background.js.